+++ docs/index.html
@@ -11,12 +11,9 @@
 </head>
 <body>
   <div id="app">
-        <header>
-          <h1>Buoy Documentation</h1>
-        </header>
-        <b-section>
-          Components
-        </b-section>
+        <b-header>
+          Buoy Documentation
+        </b-header>
         <section>
           <nav>
             <strong>Components</strong>
