+++ src/components/BHeader/BHeader.test.js
@@ -1,10 +1,10 @@
 import { mount } from 'vue-test-utils'
-import BSection from './BSection'
+import BHeader from './BHeader'
 
-describe('BSection', () => {
+describe('BHeader', () => {
   it('renders correctly', () => {
     const slot = `<div />`
-    const wrapper = mount(BSection, {
+    const wrapper = mount(BHeader, {
       slots: { default: slot }
     })
     expect(wrapper.html()).toMatchSnapshot()
