+++ src/components/BHeader/index.js
@@ -1,0 +1,8 @@
+import BHeader from './BHeader.vue'
+
+/* istanbul ignore next */
+BHeader.install = function install (Vue) {
+  Vue.component(BHeader.name, BHeader)
+}
+
+export default BHeader
