--- src/components/BSection/index.js
@@ -1,8 +1,0 @@
-import BSection from './BSection.vue'
-
-/* istanbul ignore next */
-BSection.install = function install (Vue) {
-  Vue.component(BSection.name, BSection)
-}
-
-export default BSection
