+++ src/components/index.js
@@ -1,2 +1,2 @@
 export { default as BButton } from './BButton'
-export { default as BSection } from './BSection'
+export { default as BHeader } from './BHeader'
